[naprobu-19] finish

+++ app/Library/Cron/Cron.php
@@ -191,12 +191,10 @@ class Cron
 		])->get();
 
 		$lastUser = User::where([
-            ['id','>',$queue->start],
-            ['id','<=',$queue->start + 50],
             ['status_id','<>',5],
             ['isNewsletter',1],
         ])->orderBy('id','DESC')->first();
-		if($queue->start + 50 > $lastUser->id){
+		if($lastUser && $queue->start + 50 > $lastUser->id){
 			$queue->delete();
 		}else{
 			$queue->start += 50;
@@ -375,12 +373,10 @@ class Cron
 
 
         $lastUser = User::where([
-            ['id','>',$queue->start],
-            ['id','<=',$queue->start + 150],
             ['email','<>',null],
             ['isHide',0],
         ])->orderBy('id','DESC')->first();
-		if($queue->start + 150 > $lastUser->id){
+		if($lastUser && $queue->start + 150 > $lastUser->id){
 			$queue->delete();
 		}else{
 			$queue->start += 150;
@@ -408,13 +404,11 @@ class Cron
 		])->whereNotIn('id',[1,2,7,8,9,11,12,13,14,15,16,17,18,43718,45645, 45765, 45766, 46648, 47336, 47355, 47356, 48566, 66942, 106249, 139119])->get();
 
 		$lastUser = User::where([
-            ['id','>',$queue->start],
-            ['id','<=',$queue->start + 1000],
             ['email','<>',null],
             ['new_form_status',false],
             ['isHide',0],
         ])->orderBy('id','DESC')->first();
-		if($queue->start + 1000 > $lastUser->id){
+		if($lastUser && $queue->start + 1000 > $lastUser->id){
 			$queue->delete();
 		}else{
 			$queue->start += 1000;
@@ -448,13 +442,11 @@ class Cron
         ])->whereNotIn('id',[1,2,7,8,9,11,12,13,14,15,16,17,18,43718,45645, 45765, 45766, 46648, 47336, 47355, 47356, 48566, 66942, 106249, 139119])->get();
 
         $lastUser = User::where([
-            ['id','>',$queue->start],
-            ['id','<=',$queue->start + 1000],
             ['email','<>',null],
             ['new_form_status',false],
             ['isHide',0],
         ])->orderBy('id','DESC')->first();
-        if($queue->start + 1000 > $lastUser->id){
+        if($lastUser && $queue->start + 1000 > $lastUser->id){
             $queue->delete();
         }else{
             $queue->start += 1000;
@@ -474,11 +466,9 @@ class Cron
 		])->get();
 
         $lastUser = User::where([
-            ['id','>',$queue->start],
-            ['id','<=',$queue->start + 150],
             ['phone','<>',null],
         ])->orderBy('id','DESC')->first();
-		if($queue->start + 150 > $lastUser->id){
+		if($lastUser && $queue->start + 150 > $lastUser->id){
 			$queue->delete();
 		}else{
 			$queue->start += 150;
